+++ docs/changelog.md
@@ -24,6 +24,7 @@ CodeChat Editor. If not, see
   - Resize large images to fit in browser.
   - Switch to new parser for Python and C/C++.
   - Correct styles so that the selection and current line are visible.
+  - Created a build tool to automate the build process.
 - v0.1.3, 2024-Nov-18:
   - Switch to using MathJax 4 beta; load MathJax in the frame it's used, per
     [MathJax issue #3309](https://github.com/mathjax/MathJax/issues/3309).